refactor(models): enforce uniform Request, Response, and DTO naming conventions across all payload schemas
