+++ src/Models/OrderConsentDTO.php
@@ -14,20 +14,11 @@ class OrderConsentDTO
 
     public function toArray()
     {
-        $result = [];
-        foreach (get_object_vars($this) as $key => $value) {
-            if ($value !== null) {
-                if (is_object($value) && method_exists($value, 'toArray')) {
-                    $result[$key] = $value->toArray();
-                } elseif (is_array($value)) {
-                    $result[$key] = array_map(function ($item) {
-                        return is_object($item) && method_exists($item, 'toArray') ? $item->toArray() : $item;
-                    }, $value);
-                } else {
-                    $result[$key] = $value;
-                }
-            }
-        }
-        return $result;
+        return array_filter([
+            'title' => $this->title,
+            'url' => $this->url,
+        ], function ($value) {
+            return $value !== null;
+        });
     }
 }
