+++ src/Models/OrderPaymentDetailRequest.php
@@ -1,14 +1,17 @@
 <?php
 namespace Tapsilat\Models;
 
-class CancelOrderDTO
+class OrderPaymentDetailRequest
 {
     public $reference_id;
+    public $conversation_id;
 
     public function __construct(
-        $reference_id = null
+        $reference_id = null,
+        $conversation_id = null
     ) {
         $this->reference_id = $reference_id;
+        $this->conversation_id = $conversation_id;
     }
 
     public function toArray()
