+++ src/Models/OrderPaymentTermDeleteRequest.php
@@ -1,38 +1,17 @@
 <?php
 namespace Tapsilat\Models;
 
-class OrderPaymentTermCreateDTO
+class OrderPaymentTermDeleteRequest
 {
     public $order_id;
     public $term_reference_id;
-    public $amount;
-    public $due_date;
-    public $term_sequence;
-    public $required;
-    public $status;
-    public $data;
-    public $paid_date;
 
     public function __construct(
         $order_id = null,
-        $term_reference_id = null,
-        $amount = null,
-        $due_date = null,
-        $term_sequence = null,
-        $required = null,
-        $status = null,
-        $data = null,
-        $paid_date = null
+        $term_reference_id = null
     ) {
         $this->order_id = $order_id;
         $this->term_reference_id = $term_reference_id;
-        $this->amount = $amount;
-        $this->due_date = $due_date;
-        $this->term_sequence = $term_sequence;
-        $this->required = $required;
-        $this->status = $status;
-        $this->data = $data;
-        $this->paid_date = $paid_date;
     }
 
     public function toArray()
