+++ src/Models/OrderPaymentTermUpdateRequest.php
@@ -1,17 +1,32 @@
 <?php
 namespace Tapsilat\Models;
 
-class OrderPaymentTermDeleteDTO
+class OrderPaymentTermUpdateRequest
 {
-    public $order_id;
     public $term_reference_id;
+    public $amount;
+    public $due_date;
+    public $paid_date;
+    public $required;
+    public $status;
+    public $term_sequence;
 
     public function __construct(
-        $order_id = null,
-        $term_reference_id = null
+        $amount = null,
+        $due_date = null,
+        $paid_date = null,
+        $required = null,
+        $status = null,
+        $term_reference_id = null,
+        $term_sequence = null
     ) {
-        $this->order_id = $order_id;
+        $this->amount = $amount;
+        $this->due_date = $due_date;
+        $this->paid_date = $paid_date;
+        $this->required = $required;
+        $this->status = $status;
         $this->term_reference_id = $term_reference_id;
+        $this->term_sequence = $term_sequence;
     }
 
     public function toArray()
