+++ src/Models/OrderRelatedReferenceRequest.php
@@ -1,17 +1,17 @@
 <?php
 namespace Tapsilat\Models;
 
-class OrderManualCallbackDTO
+class OrderRelatedReferenceRequest
 {
     public $reference_id;
-    public $conversation_id;
+    public $related_reference_id;
 
     public function __construct(
         $reference_id = null,
-        $conversation_id = null
+        $related_reference_id = null
     ) {
         $this->reference_id = $reference_id;
-        $this->conversation_id = $conversation_id;
+        $this->related_reference_id = $related_reference_id;
     }
 
     public function toArray()
